+++ src/app/core/service/index.ts
@@ -1,6 +1,7 @@
 export * from './app-translate.service';
 export * from './app.service';
 export * from './browser.service';
+export * from './cache.service';
 export * from './game.service';
 export * from './logger.service';
 export * from './renderer.service';