+++ src/app/shared/module/poe/provider/currency/currency-chaos-equivalents.provider.ts
@@ -1,46 +1,35 @@
 import { Injectable } from '@angular/core';
+import { CacheService } from '@app/service';
 import * as PoENinja from '@data/poe-ninja';
 import { CurrencyChaosEquivalents } from '@shared/module/poe/type';
-import { Observable, of, timer } from 'rxjs';
-import { flatMap, shareReplay, switchMap, take } from 'rxjs/operators';
+import { Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
 
-const REFRESH_INTERVAL = 1000 * 60 * 30;
-const CACHE_SIZE = 1;
+const CACHE_EXPIRY = 1000 * 60 * 30;
 
 @Injectable({
     providedIn: 'root'
 })
 export class CurrencyChaosEquivalentsProvider {
-    private readonly equivalentsPerLeague: {
-        [leagueId: string]: Observable<CurrencyChaosEquivalents>
-    } = {};
 
     constructor(
-        private readonly currencyOverviewHttpService: PoENinja.CurrencyOverviewHttpService) { }
+        private readonly currencyOverviewHttpService: PoENinja.CurrencyOverviewHttpService,
+        private readonly cache: CacheService) { }
 
     public provide(leagueId: string): Observable<CurrencyChaosEquivalents> {
-        if (!this.equivalentsPerLeague[leagueId]) {
-            const timer$ = timer(0, REFRESH_INTERVAL);
-
-            this.equivalentsPerLeague[leagueId] = timer$.pipe(
-                switchMap(_ => this.fetch(leagueId)),
-                shareReplay(CACHE_SIZE),
-            );
-        }
-        return this.equivalentsPerLeague[leagueId].pipe(
-            take(1)
-        );
+        const key = `currency_chaos_equivalents_${leagueId}`;
+        return this.cache.proxy(key, () => this.fetch(leagueId), CACHE_EXPIRY);
     }
 
     private fetch(leagueId: string): Observable<CurrencyChaosEquivalents> {
         return this.currencyOverviewHttpService.get(leagueId, PoENinja.CurrencyOverviewType.Currency).pipe(
-            flatMap(response => {
+            map(response => {
                 const currencyChaosEquivalents: CurrencyChaosEquivalents = {};
                 response.lines.forEach(line => {
                     currencyChaosEquivalents[line.currencyTypeName] = line.chaosEquivalent;
                 });
                 currencyChaosEquivalents['Chaos Orb'] = 1;
-                return of(currencyChaosEquivalents);
+                return currencyChaosEquivalents;
             })
         );
     }