+++ src/app/shared/module/poe/provider/leagues.provider.ts
@@ -1,35 +1,34 @@
 import { Injectable } from '@angular/core';
+import { CacheService } from '@app/service';
 import * as PoE from '@data/poe';
-import { Language, LanguageMap, League } from '@shared/module/poe/type';
+import { Language, League } from '@shared/module/poe/type';
 import { Observable } from 'rxjs';
-import { map, shareReplay } from 'rxjs/operators';
+import { map } from 'rxjs/operators';
 
-const CACHE_SIZE = 1;
+const CACHE_EXPIRY = 1000 * 60 * 60;
 
 @Injectable({
     providedIn: 'root'
 })
 export class LeaguesProvider {
-    private readonly languageMap: LanguageMap<Observable<League[]>> = {};
-
-    constructor(private readonly tradeHttpService: PoE.TradeHttpService) { }
+    constructor(
+        private readonly tradeHttpService: PoE.TradeHttpService,
+        private readonly cache: CacheService) { }
 
     public provide(language: Language): Observable<League[]> {
-        return this.languageMap[language] || (
-            this.languageMap[language] = this.fetch(language).pipe(shareReplay(CACHE_SIZE))
-        );
+        const key = `leagues_${language}`;
+        return this.cache.proxy(key, () => this.fetch(language), CACHE_EXPIRY);
     }
 
     private fetch(language: Language): Observable<League[]> {
-        return this.tradeHttpService.getLeagues(language)
-            .pipe(
-                map(response => response.result.map(league => {
-                    const result: League = {
-                        id: league.id,
-                        text: league.text
-                    };
-                    return result;
-                })),
-            );
+        return this.tradeHttpService.getLeagues(language).pipe(
+            map(response => response.result.map(league => {
+                const result: League = {
+                    id: league.id,
+                    text: league.text
+                };
+                return result;
+            })),
+        );
     }
 }